Add popup dialog for delete

+++ frontend/src/components/WorkoutDetails.js
@@ -3,6 +3,8 @@ import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const WorkoutTemplate = ({ workout, workouts, setNewWorkouts }) => {
   const handleDelete = async () => {
+    if (!window.confirm("Are you sure you want to delete this workout?")) return;
+
     const response = await fetch(`/api/workouts/${workout._id}`, {
       method: "DELETE",
     });
